LPS-65492 Move to the standard directory layout
